10/31

+++ MySQLSep16/Models/CarModel.cs
@@ -8,6 +8,6 @@ namespace MySQLSep16.Models
 {
     internal class CarModel
     {
-        public int engineProp { get; set; }
+        public double engineProp { get; set; }
     }
 }
